refactor(services/ghac): move raw http calls to core (#6095)
* refactor(services/ghac): move raw http calls to core

* chore: trigger CI again
